+++ packages/terra-core-docs/src/terra-dev-site/doc/section-header/example/OpenSectionHeader.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SectionHeaderExampleTemplate from 'terra-section-header/lib/terra-dev-site/doc/example/SectionHeaderExampleTemplate';
+import SectionHeaderExampleTemplate from './SectionHeaderExampleTemplate';
 
 const sectionHeaderProps = {
   title: 'Open',