+++ packages/terra-core-docs/src/terra-dev-site/test/section-header/TransparentSectionHeader.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
-import SectionHeader from '../../../SectionHeader';
+import SectionHeader from 'terra-section-header';
 
 export default () => <SectionHeader title="Transparent Section Header" isTransparent />;